oracle_python/requirements.txt
near-api-py==0.1.0